% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atak_drapieznika.R
\name{atak_drapieznika}
\alias{atak_drapieznika}
\title{Przeprowadznie atakow drapieznikow}
\usage{
atak_drapieznika(czy_lis = FALSE, czy_wilk = FALSE, stado, stado_max)
}
\arguments{
\item{czy_lis}{wartosc logiczna: TRUE - pojawil sie lis, FALSE - nie pojawil sie}

\item{czy_wilk}{wartosc logiczna: TRUE - pojawil sie wilk, FALSE - nie pojawil sie}

\item{stado}{stan stada gracza}

\item{stado_max}{maksymalny stan stada gracza}
}
\description{
Przeprowadznie atakow drapieznikow
}
\examples{
stado <- c("krolik"=3,"owca"=1,"swinia"=3,"krowa"=0,"kon"=0,"maly_pies"=0,"duzy_pies"=1)
atak_drapieznika(FALSE,FALSE,stado,stado_max)
stado_max <- c(60,24,20,12,6,4,2)
names(stado_max) <- c("krolik","owca","swinia","krowa","kon","maly_pies","duzy_pies")
stado_gracza <- c("krolik"=3,"owca"=1,"swinia"=3,"krowa"=0,"kon"=0,"maly_pies"=0,"duzy_pies"=1)
atak_drapieznika(TRUE,FALSE,stado_gracza,stado_max)
}

