% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badaj_gre.R
\name{badaj_gre}
\alias{badaj_gre}
\title{Funkcja powtarzajaca gre 10000 razy}
\usage{
badaj_gre(strategia)
}
\arguments{
\item{strategia}{strategia ktora chcemy grac}
}
\value{
podstawowe statystki dla czasow trwania gier
}
\description{
Funkcja powtarzajaca gre 10000 razy
}

