% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_zostawia_lis.R
\docType{data}
\name{co_zostawia_lis}
\alias{co_zostawia_lis}
\title{Co po ataku maksymalnie zostawia lis w stadzie gracza}
\format{nazwany wektor 7 elementowy}
\description{
Co po ataku maksymalnie zostawia lis w stadzie gracza
}

