% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_zostawia_wilk.R
\docType{data}
\name{co_zostawia_wilk}
\alias{co_zostawia_wilk}
\title{Co po ataku maksymalnie zostawia wilk w stadzie gracza}
\format{nazwany wektor 7 elementowy}
\description{
Co po ataku maksymalnie zostawia wilk w stadzie gracza
}

