% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dodaj_lub_odejmij_zwierzeta.R
\name{dodaj_lub_odejmij_zwierzeta}
\alias{dodaj_lub_odejmij_zwierzeta}
\title{Funkcja zmieniajaca stan stad: gracza i glownego}
\usage{
dodaj_lub_odejmij_zwierzeta(co, jak = 1, stado, stado_glowne)
}
\arguments{
\item{co}{mozliwe zmiany w stadzie}

\item{jak}{1-dodawanie do stada gracza, -1 odejmowanie od stada gracza}

\item{stado}{stan stada gracza}

\item{stado_glowne}{stan stada glownego}
}
\description{
Funkcja zmieniajaca stan stad: gracza i glownego
}

