% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dokonanie_wymian.R
\name{dokonanie_wymian}
\alias{dokonanie_wymian}
\title{Dokonanie wymian zwiazanych ze strategia gracza}
\usage{
dokonanie_wymian(stado, stado_max, tabela)
}
\arguments{
\item{stado}{stado gracza}

\item{stado_max}{maksymalny stan stada glownego}

\item{tabela}{tabela zmian dostepnych po wymianie}
}
\description{
Funkcja dokonanie_wymian() zwraca nam stany stad po wymianie
}

