% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jedno_na_kilka2.R
\name{jedno_na_kilka_2}
\alias{jedno_na_kilka_2}
\title{Zamiana jednego zwierzecia na kilka innych}
\usage{
jedno_na_kilka_2(wartosc, stado, co, na_co)
}
\arguments{
\item{wartosc}{wartosc zwierzat wyrazona w krolikach}

\item{stado}{stado gracza}

\item{co}{zwierze ktore chcemy oddac}

\item{na_co}{zwierze najbardziej pożądane przez gracza}
}
\value{
tabela_zmian
}
\description{
Zamieniamy jedno zwierze na kilka innych, jednemu ze zwierzat pożądanych nadajemy pewien priorytet - chcemy mieć jedną sztukę tego zwierzęcia, z reszty dostepnych srodkow kupujemy inne zwierzeta
}
\examples{
stado <- c("krolik"=3,"owca"=1,"swinia"=3,"krowa"=0,"kon"=0,"maly_pies"=0,"duzy_pies"=1)
wartosc_w_krolikach <- c(1,6,12,36,72,6,36)
names(wartosc_w_krolikach) <- c("krolik","owca","swinia","krowa","kon","maly_pies","duzy_pies")
jedno_na_kilka_2(wartosc_w_krolikach,stado,"krowa","owca")

}

