% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kilka_na_jedno.R
\name{kilka_na_jedno}
\alias{kilka_na_jedno}
\title{Funkcja wymian kilku zwierzat na jedno}
\usage{
kilka_na_jedno(wartosc, tabela_co, na_co)
}
\arguments{
\item{wartosc}{wartosc zwierzat wyrazona w krolikach}

\item{tabela_co}{tabela ze zwierzetami ktore chcemy oddac}

\item{na_co}{nazwa zwierzecia ktore chcemy dostac}
}
\description{
Funkcja kilka_na_jedno() otrzymuje nazwany wektor zawierajacy liczbe zwierzat ktAlre chcemy wymienić oraz nazwe zwierzecia ktore chcemy otrzymac
}
\examples{
tabela_co <-c("krolik"=2,"owca"=1,"swinia"=1,"krowa"=0,"kon"=0,"maly_pies"=0,"duzy_pies"=0)
kilka_na_jedno(wartosc_w_krolikach,tabela_co,"krowa")

}

