% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przebieg_gry.R
\name{przegieg_gry}
\alias{przegieg_gry}
\title{Funkcja opisujaca przebieg gry}
\usage{
przegieg_gry(strategia)
}
\arguments{
\item{strategia}{nazwa funkcji opisującej strategię gracza}
}
\description{
Funkcja opisujaca przebieg gry
}

