% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzut_kostkami.R
\name{rzut_kostkami}
\alias{rzut_kostkami}
\title{Wynik rzutu dwoma kostkami}
\usage{
rzut_kostkami(a, b)
}
\arguments{
\item{a}{wektor zawierajacy zwierzeta na kostce zielonej}

\item{b}{wektor zawierajacy zwierzeta na kostce czerwonej}
}
\value{
Wektor zawierajacy nazwy zwierzat wylosowanych na kostce zielonej i czerwonej odpowiednio
}
\description{
Funkcja rzut_kostkami() wyznacza parę wartosci - nazwy zwierzat na obu kostkach.
}
\examples{
kostka_czerwona <- c("krolik","krolik","krolik","krolik", "krolik","krolik","owca","owca","swinia","swinia","kon","lis")
kostka_zielona <- c("krolik","krolik","krolik","krolik", "krolik","krolik","owca","owca","owca","swinia","krowa","wilk")
rzut_kostkami(kostka_zielona,kostka_czerwona)

}

