% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stado_w_krolikach.R
\name{stado_w_krolikach}
\alias{stado_w_krolikach}
\title{Funkcja obliczajaca wartosc naszego stada}
\usage{
stado_w_krolikach(wartosc_w_krolikach, tabela)
}
\arguments{
\item{wartosc_w_krolikach}{wartosc danych zwierzat}

\item{tabela}{tabela dla ktorej chcemy liczyc wartosc}
}
\value{
saldo czyli wartosc tabeli
}
\description{
Funkcja obliczajaca wartosc naszego stada
}
\examples{
tabela <- c("krolik"=2,"owca"=1,"swinia"=1,"krowa"=0,"kon"=0,"maly_pies"=0,"duzy_pies"=1)
wartosc_w_krolikach <- c(1,6,12,36,72,6,36)
names(wartosc_w_krolikach) <- c("krolik","owca","swinia","krowa","kon","maly_pies","duzy_pies")
stado_w_krolikach(wartosc_w_krolikach,tabela)

}

