% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_nowa.R
\name{strategia_nowa}
\alias{strategia_nowa}
\title{Funkcja opisująca strategię gracza}
\usage{
strategia_nowa(stado)
}
\arguments{
\item{stado}{stan stada gracza}
}
\description{
Funkcja opisująca strategię gracza
}

