% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_owce.R
\name{strategia_owce}
\alias{strategia_owce}
\title{Strategia wymian z zalozeniem ze zbieramy owce}
\usage{
strategia_owce(stado)
}
\arguments{
\item{stado}{stado gracza}
}
\value{
wymiany ze strategii
}
\description{
Strategia wymian z zalozeniem ze zbieramy owce
}

