% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_wymian_0_0_0_0.R
\name{strategia_wymian_0_0_0_0}
\alias{strategia_wymian_0_0_0_0}
\title{Strategia wymian z zalozeniem ze gramy va banque}
\usage{
strategia_wymian_0_0_0_0(stado)
}
\arguments{
\item{stado}{stado gracza}
}
\value{
wymiany ze strategii
}
\description{
Strategia wymian z zalozeniem ze gramy va banque
}

