% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_wymian_1_1_1_1.R
\name{strategia_wymian_1_1_1_1}
\alias{strategia_wymian_1_1_1_1}
\title{Strategia wymian z zalozeniem ze zawsze chcemy miec jednego krolika, jedna owce, jednego malego psa i jednego duzego psa}
\usage{
strategia_wymian_1_1_1_1(stado)
}
\arguments{
\item{stado}{stado gracza}
}
\value{
wymiany ze strategii
}
\description{
Strategia wymian z zalozeniem ze zawsze chcemy miec jednego krolika, jedna owce, jednego malego psa i jednego duzego psa
}

