% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wartosc.R
\name{wartosc}
\alias{wartosc}
\title{Wartosc danego zwierzecia wyrazona w krolikach, tzn ile krolikow musimy dac aby otrzymac dane zwierze}
\usage{
wartosc(wartosc_w_krolikach, zwierze)
}
\arguments{
\item{wartosc_w_krolikach}{Wartosc w przeliczeniu na kroliki}

\item{zwierze}{Nazwa zwierzecia ktorego wartosc obliczamy}
}
\value{
wartosc zwierzecia
}
\description{
Wartosc danego zwierzecia wyrazona w krolikach, tzn ile krolikow musimy dac aby otrzymac dane zwierze
}

