% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wygrana.R
\name{wygrana}
\alias{wygrana}
\title{Funkcja sprawdzajaca czy spelnione są warunki wygranej}
\usage{
wygrana(warunek_wygranej, table)
}
\arguments{
\item{warunek_wygranej}{wektor z warunkami zakonczenia gry}

\item{table}{stan stada gracza}
}
\description{
Funkcja wygrana() przyjmuje wektor warunki wygranej oraz tabele zawierajaca stan stada gracza
}

