% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zamien_wynik_rzutu_na_wektor.R
\name{zamien_wynik_rzutu_na_wektor}
\alias{zamien_wynik_rzutu_na_wektor}
\title{Zamieniamy wynik rzutu na wektor}
\usage{
zamien_wynik_rzutu_na_wektor(wynik)
}
\arguments{
\item{wynik}{Wynik rzutu}
}
\description{
Zamieniamy wynik rzutu na wektor
}

