% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atak_drapieznika.R
\name{atak_drapieznika}
\alias{atak_drapieznika}
\title{Przeprowadznie atakow drapieznikow}
\usage{
atak_drapieznika(czy_lis = FALSE, czy_wilk = FALSE, stado, stado_max)
}
\arguments{
\item{czy_lis}{wartosc logiczna: TRUE - pojawil sie lis, FALSE - nie pojawil sie}

\item{czy_wilk}{wartosc logiczna: TRUE - pojawil sie wilk, FALSE - nie pojawil sie}

\item{stado}{stan stada gracza}

\item{stado_max}{maksymalny stan stada gracza}
}
\description{
Przeprowadznie atakow drapieznikow
}

