% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atak_drapieznika.R
\name{atak_drapieznika}
\alias{atak_drapieznika}
\title{Przeprowadznie atakow drapieznikow}
\usage{
atak_drapieznika(lis = FALSE, wilk = FALSE, stado, stado_glowne)
}
\arguments{
\item{lis}{wartosc logiczna: TRUE - pojawil sie lis, FALSE - nie pojawil sie}

\item{wilk}{wartosc logiczna: TRUE - pojawil sie wilk, FALSE - nie pojawil sie}

\item{stado}{stan stada gracza}

\item{stado_glowne}{stan stada glownego}
}
\description{
Przeprowadznie atakow drapieznikow
}

