% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badaj_gre.R
\name{badaj_gre}
\alias{badaj_gre}
\title{Funkcja powtarzajaca gre podana liczbę razy}
\usage{
badaj_gre(strategia, powtorzenia = 100,
  stado_max = SuperFarmer.SuperDziewczyn::stado_max,
  kostka_zielona = SuperFarmer.SuperDziewczyn::kostka_zielona,
  kostka_czerwona = SuperFarmer.SuperDziewczyn::kostka_czerwona,
  warunek_wygranej = SuperFarmer.SuperDziewczyn::warunek_wygranej)
}
\arguments{
\item{strategia}{strategia ktora chcemy grac}

\item{powtorzenia}{liczba powtorzen gier do stestowania, domyslnie 100}
}
\value{
wektor zawierajacy dane dotyczace dlugosci trwania wybranej liczby gier
}
\description{
Funkcja powtarzajaca gre podana liczbę razy
}

