% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badaj_gre.R
\name{badaj_gre}
\alias{badaj_gre}
\title{Funkcja powtarzajaca gre 10000 razy}
\usage{
badaj_gre(strategia, powtorzenia = 100)
}
\arguments{
\item{strategia}{strategia ktora chcemy grac}

\item{powtorzenia}{liczba powtorzen gier do stestowania, domyslnie 10000}
}
\value{
lista dwuelementowa zawierajaca podstawowe statystyki i wykres rozkladu czasow gry przy zadanej liczbie powtorzen
}
\description{
Funkcja powtarzajaca gre 10000 razy
}

