% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badaj_gre.R
\name{badaj_gre}
\alias{badaj_gre}
\title{Funkcja powtarzajaca gre podana liczbę razy}
\usage{
badaj_gre(strategia, powtorzenia = 100)
}
\arguments{
\item{strategia}{strategia ktora chcemy grac}

\item{powtorzenia}{liczba powtorzen gier do stestowania, domyslnie 100}
}
\value{
wektor zawierajacy dane dotyczace dlugosci trwania wybranej liczby gier
}
\description{
Funkcja powtarzajaca gre podana liczbę razy
}

