% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badaj_gre.R
\name{badaj_gre}
\alias{badaj_gre}
\title{Funkcja powtarzajaca gre podana liczbe razy}
\usage{
badaj_gre(strategia, powtorzenia = 100,
  stado_max = SuperFarmer.SuperDziewczyn::stado_max,
  kostka_zielona = SuperFarmer.SuperDziewczyn::kostka_zielona,
  kostka_czerwona = SuperFarmer.SuperDziewczyn::kostka_czerwona,
  warunek_wygranej = SuperFarmer.SuperDziewczyn::warunek_wygranej,
  co_zostawia_lis = SuperFarmer.SuperDziewczyn::co_zostawia_lis,
  co_zostawia_wilk = SuperFarmer.SuperDziewczyn::co_zostawia_wilk,
  wartosc_w_krolikach = SuperFarmer.SuperDziewczyn::wartosc_w_krolikach)
}
\arguments{
\item{strategia}{strategia ktora chcemy grac}

\item{powtorzenia}{liczba powtorzen gier do stestowania, domyslnie 100}

\item{stado_max}{maksymalny stan stada glownego}

\item{kostka_zielona}{zwierzeta na kostce zielonej}

\item{kostka_czerwona}{zwierzeta na kostce czerwonej}

\item{warunek_wygranej}{warunki konieczne do zakonczenia gry}

\item{co_zostawia_lis}{wektor zawierajacy liczbe zwierzat ktora maksymalnie zostaje w stadzie po ataku lisa}

\item{co_zostawia_wilk}{wektor zawierajacy liczbe zwierzat ktora maksymalnie zostaje w stadzie po ataku wilka}

\item{wartosc_w_krolikach}{wektor zawierajacy wartosci w krolikach poszczegolnych zwierzat}
}
\value{
wektor zawierajacy dane dotyczace dlugosci trwania wybranej liczby gier
}
\description{
Funkcja badaj_gre() symuluje n rozgrywek w SuperFarmera zadana na poczatku strategia. Opcjonalnie mozna zmieniac parametry, np. powtorzenia, zwierzeta na kostkach, warunek wygranej. Funkcja zwraca wektor dlugosci n zawierajacy informacje o dlugosci poszczegolnych gier.
}
\examples{
badaj_gre(strategia_owce)
warunki <- c('krolik'=3,"owca"=1,"swinia"=0,"krowa"=0,"kon"=0,"maly_pies"=0,"duzy_pies"=0)
badaj_gre(strategia_owce,warunek_wygranej=warunki)
}

