% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dostaniemy.R
\name{dostaniemy}
\alias{dostaniemy}
\title{Funkcja wyznaczajaca ile zwierzat mozemy otrzymac ze stada}
\usage{
dostaniemy(tyle_chce_wziac, stado, stado_glowne)
}
\arguments{
\item{tyle_chce_wziac}{Te zwierzeta chcemy otrzymac}

\item{stado}{Stan stada gracza}

\item{stado_glowne}{Stan stada glownego}
}
\description{
Funkcja wyznaczajaca ile zwierzat mozemy otrzymac ze stada
}

