% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dostaniemy.R
\name{dostaniemy}
\alias{dostaniemy}
\title{Funkcja wyznaczajaca ile zwierzat mozemy otrzymac ze stada}
\usage{
dostaniemy(tyle_chce_wziac, stado, stado_glowne)
}
\arguments{
\item{tyle_chce_wziac}{Te zwierzeta chcemy otrzymac}

\item{stado}{Stan stada gracza}

\item{stado_glowne}{Stan stada glownego}
}
\description{
Funkcja wyznaczajaca ile zwierzat mozemy otrzymac ze stada
}
\examples{
chcemy <- c(1,1,1,0,0,0,0)
stado <- c("krolik"=3,"owca"=1,"swinia"=3,"krowa"=0,"kon"=0,"maly_pies"=0,"duzy_pies"=1)
stado_max <- c(60,24,20,12,6,4,2)
names(stado_max) <- c("krolik","owca","swinia","krowa","kon","maly_pies","duzy_pies")
dostaniemy(chcemy,stado,stado_max-stado)

}

