% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gra.R
\name{gra}
\alias{gra}
\title{Funkcja opisująca przebieg gry}
\usage{
gra(strategia)
}
\arguments{
\item{strategia}{nazwa funkcji opisującej strategię gracza}
}
\description{
Funkcja opisująca przebieg gry
}
\examples{
gra(strategia_nowa)

}

