% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gra.R
\name{gra}
\alias{gra}
\title{Funkcja opisujaca przebieg gry}
\usage{
gra(strategia, stado_max = SuperFarmer.SuperDziewczyn::stado_max,
  kostka_zielona = SuperFarmer.SuperDziewczyn::kostka_zielona,
  kostka_czerwona = SuperFarmer.SuperDziewczyn::kostka_czerwona,
  warunek_wygranej = SuperFarmer.SuperDziewczyn::warunek_wygranej,
  co_zostawia_lis = SuperFarmer.SuperDziewczyn::co_zostawia_lis,
  co_zostawia_wilk = SuperFarmer.SuperDziewczyn::co_zostawia_wilk,
  wartosc_w_krolikach = SuperFarmer.SuperDziewczyn::wartosc_w_krolikach)
}
\arguments{
\item{strategia}{nazwa funkcji opisujacej strategie gracza}

\item{stado_max}{maksymalny stan stada glownego}

\item{kostka_zielona}{zwierzeta na kostce zielonej}

\item{kostka_czerwona}{zwierzeta na kostce czerwonej}

\item{warunek_wygranej}{warunki konieczne do zakonczenia gry}

\item{co_zostawia_lis}{wektor zawierajacy liczbe zwierzat ktora maksymalnie zostaje w stadzie po ataku lisa}

\item{co_zostawia_wilk}{wektor zawierajacy liczbe zwierzat ktora maksymalnie zostaje w stadzie po ataku wilka}

\item{wartosc_w_krolikach}{wektor zawierajacy wartosci w krolikach poszczegolnych zwierzat}
}
\value{
liczba oznaczajaca dlugosc trwania gry
}
\description{
Funkcja `gra()` symuluje pojedyncza rozgrywke w SuperFarmera zadana na poczatku strategia. Opcjonalnie mozna zmieniac parametry, np. zwierzeta na kostkach, warunek wygranej. Funkcja zwraca liczbe oznaczajaca dlugosc trwania tejze gry.
}

