% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jedno_na_kilka.R
\name{jedno_na_kilka}
\alias{jedno_na_kilka}
\title{Zamiana jednego zwierzęcia na kilka}
\usage{
jedno_na_kilka(wartosc, stado, co, na_co)
}
\arguments{
\item{wartosc}{wartosc zwierzat w krolikach}

\item{stado}{stan stada gracza}

\item{co}{nazwa zwierzęcia które chcemy oddać}

\item{na_co}{nazwa zwierzęcia które chcemy otrzymać}
}
\value{
tabela zmian, które na daną chwilę można wykonać
}
\description{
Funkcja jedno_na_kilka() zwraca nam tabele zmian które można wykonać przy zaproponowanej wymianie
}

