% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jedno_na_kilka2.R
\name{jedno_na_kilka_2}
\alias{jedno_na_kilka_2}
\title{Zamiana jednego zwierzecia na kilka innych}
\usage{
jedno_na_kilka_2(wartosc, stado, co, na_co)
}
\arguments{
\item{wartosc}{wartosc zwierzat wyrazona w krolikach}

\item{stado}{stado gracza}

\item{co}{zwierze ktore chcemy oddac}

\item{na_co}{zwierze najbardziej poLLądane przez gracza}
}
\value{
tabela_zmian
}
\description{
Zamieniamy jedno zwierze na kilka innych, jednemu ze zwierzat poLLądanych nadajemy pewien priorytet - chcemy mieć jedną sztukę tego zwierzęcia, z reszty dostepnych srodkow kupujemy inne zwierzeta
}
\examples{
jedno_na_kilka2(wartosc_w_krolikach,stado,"krowa","owca")

}

