% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kilka_na_jedno.R
\name{kilka_na_jedno}
\alias{kilka_na_jedno}
\title{Funkcja wymian kilku zwierzat na jedno}
\usage{
kilka_na_jedno(wartosc, tabela_co, na_co)
}
\arguments{
\item{wartosc}{wartosc zwierzat wyrazona w krolikach}

\item{tabela_co}{tabela ze zwierzetami ktore chcemy oddac}

\item{na_co}{nazwa zwierze ktore chcemy dostac}
}
\description{
Funkcja kilka_na_jedno() otrzymuje nazwany wektor zawierajacy liczbe zwierzat ktAlre chcemy wymienić oraz nazwe zwierzecia ktore chcemy otrzymac
}

