% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przebieg_rzutu.R
\name{przebieg_rzutu}
\alias{przebieg_rzutu}
\title{Funkcja opisujaca przebieg rzutu i zwiazane z nim zmiany w stadzie gracza i stadzie glownym}
\usage{
przebieg_rzutu(stado, stado_glowne)
}
\arguments{
\item{stado}{stan stada gracza}

\item{stado_glowne}{stan stada glownego}
}
\description{
Funkcja opisujaca przebieg rzutu i zwiazane z nim zmiany w stadzie gracza i stadzie glownym
}

