% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stado_w_krolikach.R
\name{stado_w_krolikach}
\alias{stado_w_krolikach}
\title{Funkcja obliczajaca wartosc naszego stada}
\usage{
stado_w_krolikach(wartosc_w_krolikach, tabela)
}
\arguments{
\item{wartosc_w_krolikach}{wartosc danych zwierzat}

\item{tabela}{tabela dla ktorej chcemy liczyc wartosc}
}
\value{
saldo czyli wartosc tabeli
}
\description{
Funkcja obliczajaca wartosc naszego stada
}

