% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_owce.R
\name{strategia_owce}
\alias{strategia_owce}
\title{Strategia wymian z zalozeniem ze zbieramy owce}
\usage{
strategia_owce(stado)
}
\arguments{
\item{stado}{stado gracza}
}
\value{
wymiany ze strategii
}
\description{
Funkcja `strategia_owce()` otrzymuje jako argument stan stada gracza. Zaklada ona ze jesli nie jest mozliwa wymiana na konia to zbieramy owce. Po zdobyciu dwoch koni, wymieniamy jednego na zwierzeta pozostalych gatunkow.
}
\seealso{
`strategia_wymian_0_0_0_0` i `strategia_wymian_1_1_1_1`
}

