% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_wymian_0_0_0_0.R
\name{strategia_wymian_0_0_0_0}
\alias{strategia_wymian_0_0_0_0}
\title{Strategia wymian}
\usage{
strategia_wymian_0_0_0_0(stado)
}
\arguments{
\item{stado}{stado gracza}
}
\value{
wymiany ze strategii
}
\description{
Funkcja `strategia_wymian_0_0_0_0()` otrzymuje jako argument stan stada gracza. Zaklada ona ze jesli nie jest mozliwa wymiana na konia to wymianiamy sie na mozliwie najwartosciowsze zwierze z listy. Po zdobyciu dwoch koni, wymieniamy jednego na zwierzeta pozostalych gatunkow.
}
\seealso{
`strategia_owce` i `strategia_wymian_1_1_1_1`
}

