% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_wymian_1_1_1_1.R
\name{strategia_wymian_1_1_1_1}
\alias{strategia_wymian_1_1_1_1}
\title{Strategia wymian}
\usage{
strategia_wymian_1_1_1_1(stado)
}
\arguments{
\item{stado}{stado gracza}
}
\value{
wymiany ze strategii
}
\description{
Funkcja `strategia_wymian_1_1_1_1()` otrzymuje jako argument stan stada gracza. Zaklada ona, ze jesli tylko to mozliwe, chcemy miec w stadzie glownym jednego krolika, jedna owce, jednego malego psa, jednego duzego psa. Po zdobyciu dwoch koni, wymieniamy jednego na zwierzeta pozostalych gatunkow.
}
\seealso{
`strategia_wymian_0_0_0_0`, `strategia_owce`
}

