% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizualizacja_gry.R
\name{wizualizacja_gry}
\alias{wizualizacja_gry}
\title{Wizualizacja n gier}
\usage{
wizualizacja_gry(wektor, typ_wykresu1 = "histogram", typ_wykresu2 = NULL)
}
\arguments{
\item{wektor}{wektor zawierajacy dane dotyczace dlugosci trwania poszczegolnych gier}

\item{typ_wykresu1}{wykres jaki chcemy uzyskac}

\item{typ_wykresu2}{opcjonalnie druga warstwa na wykresie}
}
\value{
wykres
}
\description{
Funkcja `wizualizacja_gry()` jako argumenty pobiera wektor dlugosci n zawierajacy informacje o dlugosci poszczegolnych gier oraz typy wykresow ktore chcemy uzyskac (mozemy wybierac sposrod : histogram, pudelkowy, skrzypcowy, pudelkowy + skrzypcowy, skrzypcowy + pudelkowy).
}

