% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wygrana.R
\name{wygrana}
\alias{wygrana}
\title{Funkcja sprawdzajaca czy spelnione są warunki wygranej}
\usage{
wygrana(warunek_wygranej, table)
}
\arguments{
\item{warunek_wygranej}{wektor z warunkami zakonczenia gry}

\item{table}{stan stada gracza}
}
\description{
Funkcja wygrana() przyjmuje wektor warunki wygranej oraz tabele zawierajaca stan stada gracza
}
\examples{
warunki_wygranej <- c("krolik"=1,"owca"=1,"swinia"=1,"krowa"=1,"kon"=1,"maly_pies"=0,"duzy_pies"=0)
stado <- c("krolik"=0,"owca"=1,"swinia"=1,"krowa"=1,"kon"=1,"maly_pies"=0,"duzy_pies"=0)
wygrana(warunki_wygranej,stado)

}

